Move common functions to a utils file
